stage1: handle docker group semantics
Docker uses the UID as GID if you only specify the "user".

In that case, docker2aci sets the group name to the user name because
the appc spec requires both user and group to be set. This will fail
because that group name won't be found in /etc/group.

Let's detect if the image was converted from docker and set the GID to
the UID in that case.