stage0,rkt: don't require the pod to be running to remove apps
The kubelet tries to remove apps from pods that are stopped and there's
no reason to prevent the removal of apps from not running pods, actually
it was considered in some parts of stage0.RmApp().

Let's not require pods to be running to remove apps. If the pod is
running, we'll still stop the app units and so on. If not, we just
remove the app.