rkt_seccomp_test: Fix arm64 stat tests
The -stat-file option of rkt's inspect program uses the standard go
os.Stat() routine, which uses the syscall.Stat() routine for its
operation.  For amd64 syscall.Stat() uses the SYS_STAT system call,
and so setting actool's --seccomp-set option to 'stat' works with
inspect's -stat-file option.

The arm64 linux kernel does not implement a SYS_STAT syscall, so for
arm64 syscall.Stat() is implemented as a wrapper around the
syscall.Fstatat() routine.  Because of this actool's --seccomp-set
option needs to be set to 'newfstatat' for the '-stat-file' option
to work correctly on arm64 machines.

Signed-off-by: Geoff Levand <geoff@infradead.org>