doc: OCI & appc/spec clarification
The documentation about OCI and appc created uncertainty for users, as
shown by the issues #3777 and #3745.

This is an attempt to clarify that although not much development is
happening in appc, rkt will not drop the ACI support.

Once this PR is merged, I would like to delete the project
https://github.com/rkt/rkt/projects/5 as it does not contain anything
and we have https://github.com/rkt/rkt/projects/4 to track the progress
of OCI native support.

Fixes https://github.com/rkt/rkt/issues/3777
Fixes https://github.com/rkt/rkt/issues/3745