Documentation: update compression note for journal files
Container Linux enabled lz4 support for journald[[1]], which means logs
started being written compressed with the lz4 format instead of xz.

Change the note about host's systemd compression support needed from xz
to lz4.

Also, clarify that this only applies to the coreos and kvm flavors.
Users of the host flavor will not be impacted since the systemd used in
stage1 will be the same as on the host.

[1]: https://github.com/coreos/coreos-overlay/pull/2593