kvm: solve routing issues by using the same default bridge as CNI
When the kvm stage1 and the coreos stage1 are used on the same machine with a
flannel CNI configuration (and likely other configurations) without a specified
bridge name, different bridges will be constructed by each stage1 for the same
subnet -- cni0 and kvm-cni0 -- causing traffic intended for the
later-constructed bridge to be lost, as it is routed to the earlier-constructed
bridge.

The normal workaround is to specify a bridge name in the CNI configuration, but
this fix avoids the issue altogether by making sure that cni0 is used in all
cases.